rename to capital Expenses; fix DOM warning
